# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger Documentation for Jobs Portal Frontend",
        "version": "1.0.0",
        "title": "Jobs Portal Frontend"
    },
    "host": "localhost:3001",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "paths": {
        "/jobs/job-roles": {
            "get": {
                "tags": [
                    "jobs"
                ],
                "summary": "Gets All Jobs",
                "description": "",
                "operationId": "job-roles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        },
...
        "/jobs/viewRoleByBand": {
            "get": {
                "tags": [
                    "jobs"
                ],
                "summary": "Gets All Role and band details to display for search",
                "description": "",
                "operationId": "viewRoleByBand",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        },
...
        "/capabilities/findByJobName": {
            "get": {
                "tags": [
                    "capabilities"
                ],
                "summary": "Get all Jobs and capabilities of roles",
                "description": "",
                "operationId": "findByJobName",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        },
...
    }
}